+++ R/DistToNearest.R
@@ -430,6 +430,13 @@ nearestDist <- function(sequences, model=c("ham", "aa", "hh_s1f", "hh_s5f", "mk_
     seq_uniq <- findUniqSeq(sequences)
     n_uniq <- length(seq_uniq)
     
+    # corresponding crossGroups values for seq_uniq
+    if (!is.null(crossGroups)) {
+        stopifnot( all.equal(sequences[match(seq_uniq, sequences)], 
+                             seq_uniq, check.attributes=FALSE) )
+        crossGroups_uniq <- crossGroups[match(seq_uniq, sequences)]
+    }
+    
     # Initialize return vector and computation vector
     seq_dist <- setNames(rep(NA, length(sequences)), sequences)
     seq_uniq_dist <- rep(NA, n_uniq)
@@ -444,6 +451,10 @@ nearestDist <- function(sequences, model=c("ham", "aa", "hh_s1f", "hh_s5f", "mk_
         # check subSampling
         subSampling <- all(!is.null(subsample), subsample < n_uniq)
         if (subSampling) indx <- sample(x=1:n_uniq, size=subsample, replace=FALSE, prob=NULL)
+        # corresponding subsampling of crossGroups_uniq
+        if (subSampling & !is.null(crossGroups)) {
+            crossGroups_uniq_sub <- crossGroups_uniq[indx]
+        }
         # Get distance matrix
         if (model == "ham") {
             if (subSampling) {
@@ -565,15 +576,57 @@ nearestDist <- function(sequences, model=c("ham", "aa", "hh_s1f", "hh_s5f", "mk_
         # Identify sequences to be considered when finding minimum
         # cross distance
         .dcross <- function(i) {
+            #cat(i,"\n")
             this_group <- crossGroups[i]
             other_groups <-  which(crossGroups != this_group)
             other_seq <- unique(sequences[other_groups])
             other_idx <- match(other_seq, seq_uniq)
             this_idx <- match(sequences[i], seq_uniq)
-            r <- dist_mat[this_idx, other_idx]
+            
+            stopifnot( all.equal( other_seq, seq_uniq[other_idx] , check.attributes=FALSE ) )
+            stopifnot( all.equal( sequences[i], seq_uniq[this_idx] , check.attributes=FALSE ) )
+            
+            # the next two checks may not always be true
+            # this happens when all the out-group sequences are identical to the in-group sequences
+            #stopifnot( all( crossGroups_uniq[other_idx] != this_group ) )
+            #stopifnot( crossGroups_uniq[this_idx] == this_group )
+            
+            if (subSampling) {
+                # When there is subsampling, nonsquareDist returns a non-n-by-n matrix 
+                # This matrix has fewers than n rows, and exactly n cols
+                # For each unique sequence, look for its cross-group distances in its column, 
+                #     NOT in its row (because there will be fewer than n rows)
+                
+                # dist_mat rows correspond to seq_uniq[indx]
+                # (indx itself is wrt seq_uniq)
+                # (other_idx is also wrt seq_uni)
+                
+                # which other_seq are included in the subsampled seqs represented by
+                #       the available rows in dist_mat?
+                # wrt dist_mat
+                other_avail_wrt_dist_mat <- which(indx %in% other_idx)
+                
+                if (length(other_avail_wrt_dist_mat)>0) {
+                    # the next two checks may not always be true
+                    # this happens when all the out-group sequences are identical to the in-group sequences
+                    #stopifnot(all( crossGroups_uniq_sub[other_avail_wrt_dist_mat] != this_group ))
+                    #stopifnot(all( crossGroups_uniq_sub[-other_avail_wrt_dist_mat] == this_group ))
+                    
+                    r <- dist_mat[other_avail_wrt_dist_mat, this_idx]
+                } else {
+                    stopifnot(all( crossGroups_uniq_sub == this_group ))
+                    return(NA)
+                }
+            } else {
+                # without subsampling
+                # dist_mat is a n-by-n matrix
+                stopifnot( all(other_idx <= nrow(dist_mat) ) ) 
+                r <- dist_mat[other_idx, this_idx]
+            }
+            
             gt0 <- which(r > 0)
             
-            if (length(gt0) != 0) { min(r[gt0]) } else { NA }
+            if (length(gt0) != 0) { return(min(r[gt0])) } else { return(NA) }
         }
         
         # Define return distance vector
@@ -586,16 +639,19 @@ nearestDist <- function(sequences, model=c("ham", "aa", "hh_s1f", "hh_s5f", "mk_
 
 #' Distance to nearest neighbor
 #'
-#' Get non-zero distance of every heavy chain sequence (as defined by \code{sequenceColumn}) 
-#' to its nearest sequence in a partition of heavy chains sharing the same V gene, J gene, 
-#' and junction length (VJL), or in a partition of single cells with heavy and light chains sharing 
-#' the same heavy chain VJL and light chain VJL combinations.
+#' Get non-zero distance of every heavy chain (\code{IGH}) sequence (as defined by 
+#' \code{sequenceColumn}) to its nearest sequence in a partition of heavy chains sharing the same 
+#' V gene, J gene, and junction length (VJL), or in a partition of single cells with heavy chains
+#' sharing the same heavy chain VJL combination, or of single cells with heavy and light chains 
+#' sharing the same heavy chain VJL and light chain VJL combinations.
 #'
 #' @param    db              data.frame containing sequence data.
-#' @param    sequenceColumn  name of the column containing heavy chain junction for grouping. 
-#'                           Note that the heavy chain junction is also used to calculate distances.
-#' @param    vCallColumn     name of the column containing the heavy chain V-segment allele calls.
-#' @param    jCallColumn     name of the column containing the heavy chain J-segment allele calls.
+#' @param    sequenceColumn  name of the column containing the junction for grouping and for calculating
+#'                           nearest neighbot distances. Note that while both heavy and light chain junctions
+#'                           may be used for VJL grouping, only the heavy chain junction is used to calculate 
+#'                           distances.
+#' @param    vCallColumn     name of the column containing the V-segment allele calls.
+#' @param    jCallColumn     name of the column containing the J-segment allele calls.
 #' @param    model           underlying SHM model, which must be one of 
 #'                           \code{c("ham", "aa", "hh_s1f", "hh_s5f", "mk_rs1nf", "hs1f_compat", "m1n_compat")}.
 #'                           See Details for further information.
@@ -611,10 +667,9 @@ nearestDist <- function(sequences, model=c("ham", "aa", "hh_s1f", "hh_s5f", "mk_
 #' @param    VJthenLen       a Boolean value specifying whether to perform partitioning as a 2-stage
 #'                           process. If \code{TRUE}, partitions are made first based on V and J
 #'                           annotations, and then further split based on junction lengths corresponding 
-#'                           to \code{sequenceColumn} (and \code{sequenceColumnLight} if specified). If
-#'                           \code{FALSE}, perform partition as a 1-stage process during which V annotation,
-#'                           J annotation, and junction length are used to create partitions simultaneously.
-#'                           Defaults to \code{TRUE}.
+#'                           to \code{sequenceColumn}. If \code{FALSE}, perform partition as a 1-stage 
+#'                           process during which V annotation, J annotation, and junction length are used 
+#'                           to create partitions simultaneously. Defaults to \code{TRUE}.
 #' @param    nproc           number of cores to distribute the function over.
 #' @param    fields          additional fields to use for grouping.
 #' @param    cross           character vector of column names to use for grouping to calculate 
@@ -622,41 +677,56 @@ nearestDist <- function(sequences, model=c("ham", "aa", "hh_s1f", "hh_s5f", "mk_
 #' @param    mst             if \code{TRUE}, return comma-separated branch lengths from minimum 
 #'                           spanning tree.
 #' @param    subsample       number of sequences to subsample for speeding up pairwise-distance-matrix calculation. 
-#'                           Subsampling is performed without replacement in each group of sequences with the 
-#'                           same \code{vCallColumn}, \code{jCallColumn}, and junction length. 
-#'                           If \code{subsample} is larger than the unique number of sequences in each group, 
-#'                           then the subsampling process is ignored for that group. For each sequence in \code{db},
-#'                           the reported \code{DIST_NEAREST} is the distance to the closest sequence in the
-#'                           subsampled set for the group. If \code{NULL} no subsampling is performed.
+#'                           Subsampling is performed without replacement in each VJL group of heavy chain sequences. 
+#'                           If \code{subsample} is larger than the unique number of heavy chain sequences in each 
+#'                           VJL group, then the subsampling process is ignored for that group. For each heavy chain
+#'                           sequence in \code{db}, the reported \code{DIST_NEAREST} is the distance to the closest
+#'                           heavy chain sequence in the subsampled set for the VJL group. If \code{NULL} no 
+#'                           subsampling is performed.
 #' @param    progress        if \code{TRUE} print a progress bar.
-#' @param    vCallColumnLight         name of the column containing the light chain V-segment allele calls. Optional.
-#' @param    jCallColumnLight         name of the column containing the light chain J-segment allele calls. Optional.                                    
-#' @param    sequenceColumnLight      name of the column containing the light chain junction for grouping. Optional. 
-#'                                    Note that the light chain junction is not used for calculating distances.
-#' @param    separator_within_seq     a single character specifying the separator between multiple annotations 
-#'                                    for a single sequence. Defaults to \code{,}.
-#' @param    separator_between_seq    a single character specifying the separator between multiple sequences. 
-#'                                    Defaults to \code{;}.
-#' @param    keepVJLgroup             a Boolean value specifying whether to keep in the output the the column 
-#'                                    column indicating grouping based on VJL combinations. Only applicable for
-#'                                    1-stage partitioning (i.e. \code{VJthenLen=FALSE}). Also see 
-#'                                    \link[alakazam]{groupGenes}.
+#' @param    cellIdColumn    name of the column containing cell IDs. Only applicable and required for 
+#'                           single-cell mode.
+#' @param    locus           name of the column containing locus information. Only applicable and 
+#'                           required for single-cell mode.
+#' @param    groupUsingOnlyIGH    use only heavy chain (\code{IGH}) sequences for VJL grouping, disregarding 
+#'                                light chains. Only applicable and required for single-cell mode. 
+#'                                Default is \code{TRUE}. Also see \link[alakazam]{groupGenes}.                   
+#' @param    keepVJLgroup         a Boolean value specifying whether to keep in the output the the column 
+#'                                column indicating grouping based on VJL combinations. Only applicable for
+#'                                1-stage partitioning (i.e. \code{VJthenLen=FALSE}). Also see 
+#'                                \link[alakazam]{groupGenes}.
 #' 
-#' @return   Returns a modified \code{db} data.frame with nearest neighbor distances in the 
-#'           \code{DIST_NEAREST} column if \code{cross=NULL}. 
-#'           if \code{cross} was specified, distances will be added as the 
-#'           \code{CROSS_DIST_NEAREST} column
+#' @return   Returns a modified \code{db} data.frame with nearest neighbor distances between heavy chain
+#'           sequences in the \code{DIST_NEAREST} column if \code{cross=NULL}. If \code{cross} was 
+#'           specified, distances will be added as the \code{CROSS_DIST_NEAREST} column. 
+#'           
+#'           Note that distances between light chain sequences are not calculated, even if light chains 
+#'           were used for VJL grouping via \code{groupUsingOnlyIGH=FALSE}. Light chain sequences, if any,
+#'           will have \code{NA} in the \code{DIST_NEAREST} field.
 #'
 #' @details
 #' 
-#' For single-cell mode, the input requirement is the same as that for \link[alakazam]{groupGenes}. 
-#' Namely, each row represents a single cell with both heavy chain and light chain V, J, and junction columns.
-#'  
-#' Note that for \code{distToNearest}, a cell/row with multiple heavy chains is not allowed.
+#' To invoke single-cell mode, both \code{cellIdColumn} and \code{locusColumn} must be supplied. 
+#' Otherwise, the function will run under non-single-cell mode.
 #' 
-#' The distance to nearest neighbor can be used to estimate a threshold for assigning Ig
-#' sequences to clonal groups. A histogram of the resulting vector is often bimodal, 
-#' with the ideal threshold being a value that separates the two modes.
+#' Under single-cell mode, only heavy chain sequences will be used for calculating nearest neighbor
+#' distances. Under non-single-cell mode, all input sequences will be used for calculating nearest
+#' neighbor distances, regardless of the values in the \code{locusColumn} field (if present).
+#' 
+#' For single-cell mode, the input format is the same as that for \link[alakazam]{groupGenes}. 
+#' Namely, each row represents a sequence/chain. Sequences/chains from the same cell are linked
+#' by a cell ID in the \code{cellIdColumn} field. Under this mode, there is a choice of whether 
+#' grouping should be done using only heavy chain (\code{IGH}) sequences only, or using both 
+#' heavy chain (\code{IGH}) and light chain (\code{IGK}, \code{IGL}) sequences. This is governed 
+#' by \code{groupUsingOnlyIGH}.
+#' 
+#' If used, values in the \code{locus} column must be one of \code{"IGH"}, \code{"IGK"}, and \code{"IGL"}.
+#' 
+#' Note that for \code{distToNearest}, a cell with multiple heavy chains is not allowed.
+#' 
+#' The distance to nearest (heavy chain) neighbor can be used to estimate a threshold for assigning 
+#' Ig sequences to clonal groups. A histogram of the resulting vector is often bimodal, with the 
+#' ideal threshold being a value that separates the two modes.
 #' 
 #' The following distance measures are accepted by the \code{model} parameter.
 #' 
@@ -679,20 +749,20 @@ nearestDist <- function(sequences, model=c("ham", "aa", "hh_s1f", "hh_s5f", "mk_
 #' }
 #' 
 #' Note on \code{NA}s: if, for a given combination of V gene, J gene, and sequence length,
-#' there is only 1 sequence (as defined by \code{sequenceColumn}), \code{NA} is returned 
-#' instead of a distance (since it has no neighbor). If for a given combination there are 
-#' multiple sequences but only 1 unique sequence, (in which case every sequence in this 
-#' group is the de facto nearest neighbor to each other, thus giving rise to distances 
+#' there is only 1  heavy chain sequence (as defined by \code{sequenceColumn}), \code{NA} is 
+#' returned instead of a distance (since it has no heavy chain neighbor). If for a given combination 
+#' there are multiple heavy chain sequences but only 1 unique one, (in which case every heavy cahin 
+#' sequence in this group is the de facto nearest neighbor to each other, thus giving rise to distances 
 #' of 0), \code{NA}s are returned instead of zero-distances.
 #' 
-#' Note on \code{subsample}: Subsampling is performed independently in each group of sequences
-#' sharing the same \code{vCallColumn}, \code{jCallColumn}, and junction length. If \code{subsample} 
-#' is larger than number of sequences in the group, it is ignored. In other words, subsampling 
-#' is performed only on groups of sequences of size equal to or greater than \code{subsample}. 
-#' \code{DIST_NEAREST} has values calculated using all sequences in the group for groups of size
-#' smaller than \code{subsample} and values calculated using a subset of sequences for the larger 
-#' groups. To select a value of \code{subsample}, it can be useful to explore the group sizes in 
-#' \code{db}.
+#' Note on \code{subsample}: Subsampling is performed independently in each VJL group for heavy chain
+#' sequences. If \code{subsample} is larger than number of heavy chain sequences in the group, it is 
+#' ignored. In other words, subsampling is performed only on groups in which the number of heavy chain 
+#' sequences is equal to or greater than \code{subsample}. \code{DIST_NEAREST} has values calculated 
+#' using all heavy chain sequences in the group for groups with fewer than \code{subsample} heavy chain
+#' sequences, and values calculated using a subset of heavy chain sequences for the larger groups. 
+#' To select a value of \code{subsample}, it can be useful to explore the group sizes in \code{db} 
+#' (and the number of heavy chain sequences in those groups).
 #' 
 #' @references
 #' \enumerate{
@@ -739,8 +809,8 @@ distToNearest <- function(db, sequenceColumn="JUNCTION", vCallColumn="V_CALL", j
                           normalize=c("len", "none"), symmetry=c("avg", "min"),
                           first=TRUE, VJthenLen=TRUE, nproc=1, fields=NULL, cross=NULL, mst=FALSE, subsample=NULL,
                           progress=FALSE,
-                          vCallColumnLight=NULL, jCallColumnLight=NULL, sequenceColumnLight=NULL, 
-                          separator_within_seq=",", separator_between_seq=";", keepVJLgroup=TRUE) {
+                          cellIdColumn=NULL, locusColumn=NULL, groupUsingOnlyIGH=TRUE, keepVJLgroup=TRUE) {
+    
     # Hack for visibility of foreach index variables
     i <- NULL
     
@@ -752,34 +822,36 @@ distToNearest <- function(db, sequenceColumn="JUNCTION", vCallColumn="V_CALL", j
     
     
     # single-cell mode?
-    # each row represents a cell with both heavy and light chains
-    if ( !is.null(vCallColumnLight) & !is.null(jCallColumnLight) & !is.null(sequenceColumnLight) ) {
-        single_cell_with_light <- TRUE
-        # check that v/j_call_light in data
-        if (!all( c(vCallColumnLight, jCallColumnLight, sequenceColumnLight) %in% colnames(db) )) {
-            stop("One or more of vCallColumnLight, jCallColumnLight, sequenceColumnLight not found as a column in db")
+    if ( !is.null(cellIdColumn) & !is.null(locusColumn) ) {
+        singleCell <- TRUE
+        
+        if (!all(db[[locusColumn]] %in% c("IGH", "IGK", "IGL"))) {
+            stop("The locus column must be one of {IGH, IGK, IGL}.")
         }
-    } else if ( is.null(vCallColumnLight) & is.null(jCallColumnLight) & is.null(sequenceColumnLight) )  {
-        single_cell_with_light <- FALSE
+        
     } else {
-        stop("If using single-cell mode, vCallColumnLight, jCallColumnLight, sequenceColumnLight must all be set")
-    }
+        singleCell <- FALSE
+    } 
     
-    columns <- c(sequenceColumn, vCallColumn, jCallColumn, fields, cross,
-                 vCallColumnLight, jCallColumnLight, sequenceColumnLight)
+    columns <- c(sequenceColumn, vCallColumn, jCallColumn, fields, cross)
     columns <- columns[!is.null(columns)]
     
     check <- checkColumns(db, columns)
     if (check != TRUE) { stop(check) }
     
     # Convert sequence columns to uppercase
-    db <- toupperColumns(db, c(sequenceColumn, sequenceColumnLight)) 
+    db <- toupperColumns(db, c(sequenceColumn)) 
     
-    # Disallow multiple heavy chains per row
-    bool <- sapply(db[[sequenceColumn]], stringi::stri_detect_fixed, pattern=separator_between_seq)
-    if (any(bool)) {
-        stop("Detected multiple heavy chains in row(s). Each row must contain only 1 heavy chain.")
+    # Disallow multiple heavy chains per cell
+    if (singleCell) {
+        bool <- sapply(unique(db[[cellIdColumn]]), function(x) { 
+            return( sum( db[[locusColumn]][db[[cellIdColumn]]==x] == "IGH" )>1 ) 
+        } )
+        if (any(bool)) {
+            stop("Detected multiple heavy chains in cell(s). Each cell must contain only 1 heavy chain.")
+        }
     }
+    
     
     # Check for invalid characters
     # heavy
@@ -791,67 +863,28 @@ distToNearest <- function(db, sequenceColumn="JUNCTION", vCallColumn="V_CALL", j
         db <- db[valid_seq, ]
     }
     
-    # light
-    if (single_cell_with_light) {
-        valid_seq <- sapply(db[[sequenceColumnLight]], function(x, separator) {
-            if (stringi::stri_detect_fixed(str=x, pattern=separator)) {
-                x_split <- stringi::stri_split_fixed(str=x, pattern=separator)[[1]]
-                all(sapply(x_split, allValidChars, getCharsInModel(model)))
-            } else {
-                allValidChars(x, getCharsInModel(model))
-            }
-        }, separator_between_seq) 
-        not_valid_seq <- which(!valid_seq)
-        if (length(not_valid_seq)>0) {
-            warning("Invalid sequence characters in the ", sequenceColumnLight, " column.",
-                    length(not_valid_seq)," sequence(s) removed")
-            db <- db[valid_seq,]
-        } 
-    }
-    
-    
-    # helper function to calc junction legnth for individual junctions
-    .getJuncLen <- function(x, separator) {
-        if (stringi::stri_detect_fixed(str=x, pattern=separator)) {
-            x_split <- stringi::stri_split_fixed(str=x, pattern=separator)[[1]]
-            x_split_len <- stri_length(x_split)
-            juncLen <- paste(x_split_len, collapse=";")
-        } else {
-            juncLen <- as.character(stri_length(x))
-        }
-        return(juncLen)
-    }
-    
     # junction length columns (prep for groupGenes)
     junc_len <- "JUNC_LEN"
-    db[[junc_len]] <- sapply(db[[sequenceColumn]], .getJuncLen, separator=separator_between_seq)
-    if (single_cell_with_light) {
-        junc_len_light <- "JUNC_LEN_LIGHT"
-        db[[junc_len_light]] <- sapply(db[[sequenceColumnLight]], .getJuncLen, separator=separator_between_seq)
-    } else {
-        junc_len_light <- NULL
-    }
+    db[[junc_len]] <- stri_length(db[[sequenceColumn]])
     
     # create V+J grouping, or V+J+L grouping
     if (VJthenLen) {
         # 2-stage partitioning using first V+J and then L
         # V+J only first
         # creates $VJ_GROUP
-        db <- groupGenes(data=db, v_call=vCallColumn, j_call=jCallColumn, junc_len=NULL,
-                         first=first, v_call_light=vCallColumnLight, j_call_light=jCallColumnLight, 
-                         junc_len_light=NULL,
-                         separator_within_seq=separator_within_seq, separator_between_seq=separator_between_seq)
+        db <- groupGenes(db, v_call=vCallColumn, j_call=jCallColumn, junc_len=NULL,
+                         cell_id=cellIdColumn, locus=locusColumn, only_igh=groupUsingOnlyIGH,
+                         first=first)
         # L (later)  
-        group_cols <- c("VJ_GROUP", junc_len, junc_len_light)
+        group_cols <- c("VJ_GROUP", junc_len)
         
     } else {
         # 1-stage partitioning using V+J+L simultaneously
         # creates $VJ_GROUP
         # note that despite the name (VJ), this is based on V+J+L
-        db <- groupGenes(data=db, v_call=vCallColumn, j_call=jCallColumn, junc_len=junc_len,
-                         first=first, v_call_light=vCallColumnLight, j_call_light=jCallColumnLight, 
-                         junc_len_light=junc_len_light,
-                         separator_within_seq=separator_within_seq, separator_between_seq=separator_between_seq)
+        db <- groupGenes(db, v_call=vCallColumn, j_call=jCallColumn, junc_len=junc_len,
+                         cell_id=cellIdColumn, locus=locusColumn, only_igh=groupUsingOnlyIGH,
+                         first=first)
         group_cols <- c("VJ_GROUP")
     }
     
@@ -920,7 +953,9 @@ distToNearest <- function(db, sequenceColumn="JUNCTION", vCallColumn="V_CALL", j
                                  "calcTargetingDistance",
                                  "findUniqSeq",
                                  "pairwise5MerDist",
-                                 "nonsquare5MerDist")
+                                 "nonsquare5MerDist",
+                                 "singleCell",
+                                 "locusColumn")
         parallel::clusterExport(cluster, export_functions, envir=environment())
     }
     
@@ -931,20 +966,33 @@ distToNearest <- function(db, sequenceColumn="JUNCTION", vCallColumn="V_CALL", j
         pb <- progressBar(n_groups) 
     }
     list_db <- foreach(i=1:n_groups, .errorhandling='stop') %dopar% {
+        # wrt db
         idx <- uniqueGroupsIdx[[i]]
+        
+        if (singleCell) {
+            # only use IGH
+            # wrt idx
+            idxBool <- db[[locusColumn]][idx] == "IGH"
+        } else {
+            idxBool <- rep(TRUE, length(idx))
+        }
+        
         db_group <- db[idx, ]
+        
         crossGroups <- NULL
         if (!is.null(cross)) {
             crossGroups <- db_group %>% dplyr::group_indices(!!!rlang::syms(cross))
         }
-        arrSeqs <-  as.vector(unlist(db[idx, sequenceColumn]))
-        db_group$TMP_DIST_NEAREST <- nearestDist(arrSeqs, 
-                                                 model=model,
-                                                 normalize=normalize,
-                                                 symmetry=symmetry,
-                                                 crossGroups=crossGroups,
-                                                 mst=mst,
-                                                 subsample=subsample)
+        
+        arrSeqs <-  db[[sequenceColumn]][idx]
+            
+        db_group$TMP_DIST_NEAREST[idxBool] <- nearestDist(arrSeqs[idxBool], 
+                                                          model=model,
+                                                          normalize=normalize,
+                                                          symmetry=symmetry,
+                                                          crossGroups=crossGroups[idxBool],
+                                                          mst=mst,
+                                                          subsample=subsample)
         # Update progress
         if (progress) { pb$tick() }
         
@@ -968,7 +1016,7 @@ distToNearest <- function(db, sequenceColumn="JUNCTION", vCallColumn="V_CALL", j
     if ((!VJthenLen) && keepVJLgroup) {
         db$VJL_GROUP <- db[["VJ_GROUP"]]
     }
-    db <- db[, !(names(db) %in% c(junc_len, junc_len_light, "VJ_GROUP", "ROW_ID", "V1", "J1","TMP_DIST_NEAREST"))]
+    db <- db[, !(names(db) %in% c(junc_len, "VJ_GROUP", "ROW_ID", "V1", "J1","TMP_DIST_NEAREST"))]
     
     return(db)
 }